// src/lib/suap/generated/models/boletimSchema.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * SUAP API
 * OpenAPI spec version: 1.0.0
 */
import type { NotaBoletimSchema } from "./notaBoletimSchema"

export interface BoletimSchema {
  codigo_diario?: string
  disciplina?: string
  segundo_semestre?: boolean
  carga_horaria?: number
  carga_horaria_cumprida?: number
  numero_faltas?: number
  percentual_carga_horaria_frequentada?: number
  situacao?: string
  quantidade_avaliacoes?: number
  nota_etapa_1?: NotaBoletimSchema
  nota_etapa_2?: NotaBoletimSchema
  nota_etapa_3?: NotaBoletimSchema
  nota_etapa_4?: NotaBoletimSchema
  media_disciplina?: number | null
  nota_avaliacao_final?: NotaBoletimSchema
  media_final_disciplina?: string | null
}